app/config/middleware.py
import uuid
import structlog
from fastapi import Request
from starlette.middleware.base import BaseHTTPMiddleware

class RequestContextMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        structlog.contextvars.clear_contextvars()
        structlog.contextvars.bind_contextvars(
            request_id=str(uuid.uuid4()),
            path=request.url.path,
            method=request.method,
            client_host=request.client.host,
        )
        response = await call_next(request)
        return response
